--- MVP/src/crdt/crdt_test.go
@@ -1,1 +1,0 @@
-package crdt_test
